requirements.txt
# Dependencies
pandas
openpyxl
google-genai
python-dotenv
Pillow
requests
streamlit
streamlit

